New expressions
Now generating code for If expressions and Function calls

Improved function/procedure declaration parameter parsing (0 parameters allowed)

Broke monad part of LLVM.State into LLVM.Monad

Added new codegen combinators in LLVM.Monad, specifically
- `(a #)`: marks the beginning of a new block with the name `a`.
- terminate: gives a Named Terminator to the current block and places it at
  the end of the list of blocks.
- terminate': same as terminate, but the Terminator isn't Named.

  The main idea here is to make the closing and opening of blocks separate,
  deliberate actions.

Restructured Program parser. A program now looks like this:
```
program <name> begin

<functions/procedures>

main |[ <instructions> ]|

<functions/procedures>

end
```
The functions/procedures after the main block won't be reachable, of course,
but it'd be nice to show this with a "function not defined" error rather than
with an ugly syntax error.

I think that's it.

+++ source/LLVM/Type.hs
@@ -39,7 +39,7 @@ boolType :: LLVM.Type
 boolType   = i1
 
 stringType :: LLVM.Type
-stringType = LLVM.PointerType i16 (LLVM.AddrSpace 0)
+stringType = LLVM.PointerType i8 (LLVM.AddrSpace 0)
 
 
 toLLVMType :: T.Type -> LLVM.Type
@@ -63,8 +63,3 @@ sizeOf T.GChar         = 4
 sizeOf T.GFloat        = 4
 sizeOf (T.GArray sz t) = (fromIntegral sz) * sizeOf t
 sizeOf (T.GPointer t)  = 4
-
-
-
-
-
